requirements.txt
asn1crypto==1.5.1
azure-common==1.1.28
azure-core==1.31.0
azure-storage-blob==12.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==41.0.7
cx-Oracle==8.3.0
filelock==3.16.1
idna==3.10
isodate==0.6.1
numpy==2.0.2
packaging==24.1
pandas==2.2.3
platformdirs==4.3.6
pyarrow==17.0.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
snowflake-connector-python==3.12.2
sortedcontainers==2.4.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
